pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "active-boxes"
description = "Tiny ActivityPub framework written in Python, both database and server agnostic."
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Chaiwat Suttipongsakul", email = "cwt@bashell.com"},
]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: Implementation :: CPython",
]
version = "0.1.0"

[project.urls]
Homepage = "https://github.com/cwt/active-boxes"

[tool.poetry]
name = "active-boxes"
version = "0.1.0a1"
description = "Tiny ActivityPub framework written in Python, both database and server agnostic."
authors = ["Chaiwat Suttipongsakul <cwt@bashell.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/cwt/active-boxes"
repository = "https://github.com/cwt/active-boxes"

[tool.poetry.dependencies]
python = ">=3.10"
requests = ">=2.31.0"
aiohttp = ">=3.9.0"
markdown = ">=3.4.0"
bleach = ">=6.0.0"
pyld = ">=2.0.0"
pycryptodome = ">=3.18.0"
html2text = ">=2020.1.16"
mdx_linkify = ">=1.5.0"
regex = ">=2023.0.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.poetry.group.dev.dependencies]
httpretty = ">=1.1.4"
black = ">=23.0.0"
codecov = ">=2.1.13"
flake8 = ">=6.0.0"
ruff = ">=0.1.0"
pytest = "^9.0"
pytest-cov = "^7.1"
mypy = "^1.17"
pytest-xdist = "^3.8.0"
pytest-asyncio = ">=0.24.0"
types-Markdown = "*"
types-regex = "*"
types-requests = "*"
